+++ src/main/java/it/pc/test/WebSpringApp/dto/ProduttoreDTO.java
@@ -1,7 +1,5 @@
 package it.pc.test.WebSpringApp.dto;
 
-import it.pc.test.WebSpringApp.entity.PartnerEntity;
-import it.pc.test.WebSpringApp.entity.SedeEntity;
 import it.pc.test.WebSpringApp.enums.TipoCategoria;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -15,15 +13,11 @@ import java.util.List;
 public class ProduttoreDTO extends AbstractBaseDTO<Integer> {
 
     private Integer id;
-
     private String nome;
-
     private String descrizione;
-
     private PartnerDTO partner;
-
     private TipoCategoria categoria;
-    private Double numeroDipendenti;
+    private Integer numeroDipendenti;
     private Double fatturatoAnnuo;
     private SedeDTO sede;
     private List<ProdottoDTO> prodotti;
